some refactoring changes
