+++ app/src/main/java/com/polant/projectsport/activity/ActivityCalculateFood.java
@@ -30,8 +30,8 @@ import com.polant.projectsport.ThemeSettings;
 import com.polant.projectsport.data.Database;
 import com.polant.projectsport.data.model.SpecificFood;
 import com.polant.projectsport.data.model.UserParametersInfo;
-import com.polant.projectsport.fragment.CalculateDetailsFoodFragment;
-import com.polant.projectsport.fragment.CalculateFoodFragment;
+import com.polant.projectsport.fragment.food.CalculateDetailsFoodFragment;
+import com.polant.projectsport.fragment.food.CalculateFoodFragment;
 import com.polant.projectsport.fragment.dialog.TodayFoodDialogFragment;
 import com.polant.projectsport.preferences.PreferencesNewActivity;
 import com.polant.projectsport.preferences.PreferencesOldActivity;
